# Conflicts:
#	client/src/main/resources/localization/ticketlineClient_de.properties
#	client/src/main/resources/localization/ticketlineClient_en.properties

+++ server/src/main/java/at/ac/tuwien/inso/sepm/ticketline/server/service/implementation/SimpleReservationService.java
@@ -1,10 +1,11 @@
 package at.ac.tuwien.inso.sepm.ticketline.server.service.implementation;
 
+import at.ac.tuwien.inso.sepm.ticketline.server.configuration.properties.PriceCalculationProperties;
 import at.ac.tuwien.inso.sepm.ticketline.server.entity.*;
-import at.ac.tuwien.inso.sepm.ticketline.server.exception.InvalidReservationException;
 import at.ac.tuwien.inso.sepm.ticketline.server.exception.service.InternalCancelationException;
 import at.ac.tuwien.inso.sepm.ticketline.server.exception.service.InternalHallValidationException;
 import at.ac.tuwien.inso.sepm.ticketline.server.exception.service.InternalSeatReservationException;
+import at.ac.tuwien.inso.sepm.ticketline.server.exception.service.InvalidReservationException;
 import at.ac.tuwien.inso.sepm.ticketline.server.repository.PerformanceRepository;
 import at.ac.tuwien.inso.sepm.ticketline.server.repository.ReservationRepository;
 import at.ac.tuwien.inso.sepm.ticketline.server.service.HallPlanService;
@@ -16,14 +17,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-import javax.transaction.Transactional;
-import javax.validation.ConstraintViolationException;
 import java.lang.invoke.MethodHandles;
 import java.time.LocalDateTime;
 import java.util.LinkedList;
 import java.util.List;
-import java.util.NoSuchElementException;
 
 @Service
 public class SimpleReservationService implements ReservationService {
@@ -39,7 +38,31 @@ public class SimpleReservationService implements ReservationService {
     @Autowired
     private SeatsService seatsService;
 
+    private PriceCalculationProperties priceCalculationProperties;
+    //   private long lastReservationNumber = 16777216;
 
+    public ReservationRepository getReservationRepository() {
+        return reservationRepository;
+    }
+
+    @Override
+    public void setReservationRepository(ReservationRepository reservationRepository) {
+        this.reservationRepository = reservationRepository;
+    }
+
+    @Override
+    public void setPerformanceRepository(PerformanceRepository performanceRepository) {
+        this.performanceRepository = performanceRepository;
+    }
+
+    @Override
+    public void setSeatsService(SeatsService seatsService) {
+        this.seatsService = seatsService;
+    }
+
+    public SimpleReservationService(PriceCalculationProperties priceCalculationProperties) {
+        this.priceCalculationProperties = priceCalculationProperties;
+    }
     @Override
     public List<Reservation> findAllByEventId(Long eventId) {
         return reservationRepository.findAllByEventId(eventId);
@@ -78,6 +101,7 @@ public class SimpleReservationService implements ReservationService {
     }
 
     @Override
+    @Transactional
     public Reservation createAndPayReservation(Reservation reservation) throws InvalidReservationException, InternalSeatReservationException {
         createReservation(reservation);
         return purchaseReservation(reservation);
@@ -89,20 +113,27 @@ public class SimpleReservationService implements ReservationService {
     }
 
     @Override
+    @Transactional
     public Reservation editReservation(Reservation reservation) throws InvalidReservationException {
         List<Seat> changedSeats = reservation.getSeats();//the changed seats
-        List<Seat> savedSeats = reservationRepository.findByPaidFalseAndId(reservation.getId()).getSeats(); //the old seat config
+        Reservation savedReservation = reservationRepository.findByPaidFalseAndId(reservation.getId());
+        List<Seat> savedSeats = savedReservation.getSeats();//the old seat config
         List<Seat> onlyNewSeats = getNewSeats(changedSeats);//the not yet saved seats
 
         //First, get the picked performance for this reservation
-        Performance performance = performanceRepository.findById(reservation.getPerformance().getId()).orElse(null);
+        List<Sector> sectors = null;
+        Performance performance = null;
+        if (performanceRepository.findById(reservation.getPerformance().getId()).isPresent()) {
+            performance = performanceRepository.findById(reservation.getPerformance().getId()).get();
+            sectors = performance.getHall().getSectors();
+        }
 
         //Then, too fail fast, we check the integrity of the seats according to the hall plan and the sectors
         try {
-            if (performance != null) {
-                hallPlanService.checkSeatsAgainstSectors(onlyNewSeats, performance.getHall().getSectors());
+            if (sectors != null) {
+                hallPlanService.checkSeatsAgainstSectors(onlyNewSeats, sectors);
             } else {
-                LOGGER.error("Could not find the the Reservation");
+                LOGGER.error("Could not find the the Sectors for this reservation");
                 throw new InvalidReservationException("The Performance was not set");
             }
         } catch (InternalHallValidationException i) {
@@ -226,50 +257,53 @@ public class SimpleReservationService implements ReservationService {
         Performance currentPerformance = performanceRepository.findById(reservation.getPerformance().getId()).get();
 
         Reservation createdReservation = null;
-        //Generate a unique ID for the reservation
-        boolean unique = false;
-        do {
-            try {
-                reservation.setReservationNumber(generateReservationNumber());
-                //When ID creation is successful, store the reservation
-                unique = true;
-            } catch (ConstraintViolationException e) {
-                unique = false;
-            }
-        } while (!unique);
+
+        reservation.setReservationNumber(generateReservationNumber());
 
         //Set all the information we need and save
         reservation.setSeats(createdSeats);
         reservation.setPerformance(currentPerformance);
+
+        Long elusivePrice = calculatePrice(reservation);
+        reservation.setElusivePrice(elusivePrice);
 
         LOGGER.debug("Storing reservation");
         createdReservation = reservationRepository.save(reservation);
         return createdReservation;
     }
 
+
+
     public String generateReservationNumber() {
-        String reservationNumber = "";
-        final String ALPHA_NUMERIC_STRING = "ABCDEFGHJKLMNPQRSTUVWXYZ0123456789";
-        while (reservationNumber.length() <= 6) {
-            int character = (int) (Math.random() * ALPHA_NUMERIC_STRING.length());
-            reservationNumber += ALPHA_NUMERIC_STRING.charAt(character);
+        String reservationNumber;
+        Reservation previous = reservationRepository.findFirstByOrderByIdDesc();
+        //if this is the first generated reservation
+        if (previous == null) {
+            //First HEX Number with 7 Numbers
+            long first = 16777216;
+            reservationNumber = Long.toHexString(first);
+            reservationNumber = reservationNumber.toUpperCase();
+        } else {
+            String lastNumber = previous.getReservationNumber();
+            Long lastNumberLong = Long.parseLong(lastNumber, 16);
+            reservationNumber = Long.toHexString(++lastNumberLong);
+            reservationNumber = reservationNumber.toUpperCase();
         }
         return reservationNumber;
     }
 
-
-
-
     @Override
     public Reservation cancelReservation(Long id) throws InternalCancelationException {
         Reservation reservation = reservationRepository.findById(id).orElseThrow(InternalCancelationException::new);
+        if (reservation.getPerformance().getPerformanceStart().isBefore(LocalDateTime.now())) {
+            throw new InternalCancelationException("The performance has already started and the booking can't be canceled!");
+        }
         List<Seat> seatsOfReservation = reservation.getSeats();
         reservation.setSeats(null);
         reservation.setCanceled(true);
-        Reservation canceledReservation = reservationRepository.save(reservation);
+        reservationRepository.save(reservation);
         seatsOfReservation.forEach(seat -> seatsService.deleteSeat(seat));
-
-        return canceledReservation;
+        return reservation;
     }
 
     @Override
@@ -279,12 +313,35 @@ public class SimpleReservationService implements ReservationService {
 
     @Override
     public Long calculatePrice(Reservation reservation) {
+        Long calculatedPrice = calculatePreTaxPrice(reservation);
+
+        calculatedPrice += Math.round(calculatedPrice * priceCalculationProperties.getSalesTax());
+        return calculatedPrice;
+    }
+
+    @Override
+    public Long calculatePreTaxPrice(Reservation reservation) {
         Long performancePrice = reservation.getPerformance().getPrice();
         Long calculatedPrice = 0l;
         for(Seat s: reservation.getSeats()) {
             calculatedPrice += performancePrice * s.getSector().getCategory().getBasePriceMod();
         }
-
         return calculatedPrice;
+    }
+
+    /**
+     * For now we simply return the whole amount of money
+     */
+    @Override
+    public Long calculateReimbursedAmount(Reservation reservation) {
+        Long calculatedPrice = calculatePreTaxPrice(reservation);
+
+        calculatedPrice += Math.round(calculatedPrice * priceCalculationProperties.getSalesTax());
+        return calculatedPrice;
+    }
+
+    @Override
+    public Double getRegularTaxRate() {
+        return priceCalculationProperties.getSalesTax();
     }
 }
